added maintenance and allowed flags for Machine
allows the server to return to the arduino client the status of the machine: if maintenance is needed, and if usage of the machine is allowed.
board now checks server at startup and tries to get up-to-date machine information.
server returns name of holder if uid is valid.
not sure if the actual usage time of the machine shall be monitored by arduino (maybe based on one RUNNING signal from the machine) or done purely server-side, based on login/logout times.