MQTT: added regular "inUse" message
Goal is to limit information loss if the machine and the board are powered off during use. With this information the backend can save the last time the machine was used and eventually close the record at next start or after delay,
Renamed FabServer to FabBackend for consistency with Python project.